+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Logistical Form : Fix the exception managed per TraceBack method when sequence is not defined.
 - Fix accounting notification process.
 - Subrogation Release : Fix the status reset and clear invoice set when copy.
+- Subrogation Release : Fix throw exception and trace it when sequence is not defined.
 
 ## [5.0.6] - 2018-10-06
 ## Features